LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/drawing-text-to-the-canvas/2d.text.drawing.style.spacing.measure-expected.txt

2d.text.drawing.style.spacing.measure
Testing letter spacing and word spacing
Actual output:

FAIL Testing letter spacing and word spacing assert_equals: ctx.letterSpacing === 0 (got [undefined], expected 0[number]) expected (number) 0 but got (undefined) undefined
